requirements.txt
asgiref==3.8.1
black==24.10.0
click==8.1.7
Django==5.0.9
django-allauth==65.1.0
mypy-extensions==1.0.0
packaging==24.1
pathspec==0.12.1
platformdirs==4.3.6
psycopg==3.1.19
psycopg-binary==3.1.19
sqlparse==0.5.1
typing_extensions==4.12.2
django-widget-tweaks==1.5.0
pillow==11.0.0
fontawesomefree==6.6.0
environs[django]==11.0.0
django-formtools==2.5.1
django-filter==24.3
django-debug-toolbar==3.4.0
whitenoise==6.1.0
gunicorn==20.1.0